Fixes necessary to make use of playlist syncing on Gpodder Service
Created a PlaylistManager::setupTemporarySync() to create synchronisations that doesn't
persist between amarok executions. Added more documentation to the synchornisation
algorithm in PlaylistManager::addPlaylist(). Fixed SqlPodcastChannel::syncTrackStatus()
that now synchronize track status with any track derived from PodcastEpisode. Created another
constructor to SqlPodcastEpisode that can link an episode with a sqlPodcastChannel passed as
parameter. Denyed the the synchronisation between a slave and a master, that have the same
provider, in SyncedPlaylist::addPlaylist().